requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
pillow>=9.5.0
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.7.0
scikit-image>=0.20.0
tqdm>=4.65.0
gradio>=3.35.0
albumentations>=1.3.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
einops>=0.6.1
facenet-pytorch>=2.5.2
lpips>=0.1.4
matplotlib>=3.7.0
pandas>=2.0.0
